Propagate rollout dtype to P2P CPU replicas
P2P builds a sharded SGLang model locally before transferring weights. Passing only the model path discarded the rollout server dtype, so float32 HF metadata silently resolved auto to fp16 even when live engines explicitly used bf16. Preserve the queried rollout dtype and pin both explicit and auto behavior in regression coverage.

+++ miles/backends/megatron_utils/update_weight/update_weight_from_distributed/p2p.py
@@ -267,7 +267,7 @@ class UpdateWeightP2P(DistBucketedWeightUpdateMixin):
         initialize_fp4_gemm_config(server_args)
         with ParallelismContext(parallelism_config):
             model = get_model(
-                model_config=ModelConfig(model_path),
+                model_config=ModelConfig(model_path, dtype=server_args.dtype),
                 load_config=load_config,
                 device_config=DeviceConfig(),
             )
